+++ core/benches/basic.rs
@@ -1,0 +1,7 @@
+#![allow(warnings)]
+
+mod units;
+
+criterion::criterion_main! {
+    units::basic_mapx_raw::benches,
+}