+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # CHANGE LOG
 
+## v0.53.x
+
+#### New APIs
+
+- Add various trie-based/mpt-based functions
+
+#### Function changes
+
+- Disable the "vs" feature by default
+
+## v0.52.x
+
+#### Function changes
+
+- Extend the feature of "derive" to "vs"
+
 ## v0.51.x
 
 #### Function changes